BAH-1503 | Refactor. rename patient lab chart component and url to patient lab details
co-authored by Priyanka
