+++ src/config-schema.ts
@@ -1,9 +1,9 @@
 import {Type, validators} from '@openmrs/esm-framework'
-import {spaRoot, patientLabChartPath} from './constants'
+import {spaRoot, patientLabDetailsPath} from './constants'
 export const configSchema = {
   search: {
     patientResultUrl: {
-      _default: spaRoot + patientLabChartPath,
+      _default: spaRoot + patientLabDetailsPath,
       _description:
         'Where clicking a patient result takes the user. Accepts template parameter ${patientUuid}',
       _validators: [validators.isUrlWithTemplateParameters(['patientUuid'])],
