+++ src/constants.ts
@@ -1,5 +1,5 @@
 export const spaRoot = window['getOpenmrsSpaBase']()
 export const labliteModuleName = '@bahmni/lab-lite-app'
 export const searchModuleName = '@openmrs/esm-patient-search-app'
-export const patientLabChartPath = 'patient/${patientUuid}/chart/lab'
-export const patientLabChartRoute = '/patient/:patientUuid/chart/lab'
+export const patientLabDetailsPath = 'patient/${patientUuid}'
+export const patientLabDetailsRoute = '/patient/:patientUuid'
