+++ src/patient-lab-details/patient-lab-details.scss
